Move all interfaces to a separate module
